Changed where images are saved and updated other.

+++ src/engine/GameObject.java
@@ -4,9 +4,9 @@ import java.awt.Graphics;
 
 public interface GameObject {
 	public void update();
-	public void draw(Graphics g, int cameraX, int cameraY);
-	public int getX();
-	public int getY();
-	public int getHeight();
-	public int getWidth();
+	public void draw(Graphics g, double cameraX, double cameraY);
+	public double getX();
+	public double getY();
+	public double getHeight();
+	public double getWidth();
 }
